+++ packages/dara-core/changelog.md
@@ -2,7 +2,7 @@
 title: Changelog
 ---
 
-## NEXT
+## 1.10.0
 
 -   Fixed an issue where in some cases a `DerivedDataVariable` would be invoked with an internal `PendingValue`
 -   Implement `Variable.init_override` static method to allow overriding how variables are initialized within a given context.